requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-comments-xtd==2.10.8
django-cors-headers==4.3.1